ajout test dans class appart
