+++ core/pom.xml
@@ -152,10 +152,12 @@
                         <targetClasses>
                             <param>com.graphhopper.config.Profile</param>
                             <param>com.graphhopper.storage.IntsRef</param>
+                            <param>com.graphhopper.GraphHopper</param>
                         </targetClasses>
                         <targetTests>
                             <param>com.graphhopper.GraphHopperProfileTest</param>
                             <param>com.graphhopper.storage.IntsRefTest</param>
+                            <param>com.graphhopper.GraphHopperTest2</param>
                         </targetTests>
                         <outputFormats>
                             <outputFormat>HTML</outputFormat>
