+++ core/src/test/java/com/graphhopper/GraphHopperTest.java
@@ -2877,73 +2877,4 @@ public class GraphHopperTest {
         assertEquals(1, p.get(1).getLast());
         assertEquals(0.0, (double) p.get(1).getValue(), 1.e-3);
     }
-
-        @Test
-        void getCountryRuleFactory() {
-
-                GraphHopper hopper = new GraphHopper();
-                CountryRuleFactory countryRuleFactory = new CountryRuleFactory();
-                
-                // set et get countryRuleFactory
-                hopper.setCountryRuleFactory(countryRuleFactory);
-                assertEquals(countryRuleFactory, hopper.getCountryRuleFactory());
-
-                // set countryRuleFactory à null
-                hopper.setCountryRuleFactory(null);
-                assertNull(hopper.getCountryRuleFactory());
-        }
-
-        @Test
-        void getImportRegistry() {
-
-                GraphHopper hopper = new GraphHopper();
-                ImportRegistry importRegistry = new DefaultImportRegistry();
-                
-                // set et get importRegistry
-                hopper.setImportRegistry(importRegistry);
-                assertEquals(importRegistry, hopper.getImportRegistry());
-        }
-
-        @Test
-        void get_ET_setCustomAreasDirectory() {
-
-                GraphHopper hopper = new GraphHopper();
-                String fakerPathString = new Faker().file().fileName(); // définir chemin aléatoire
-
-                // set et get avec chemin aléatoire
-                hopper.setCustomAreasDirectory(fakerPathString);
-                assertEquals(fakerPathString, hopper.getCustomAreasDirectory());
-        }
-
-        @Test
-        void setLocationIndexThrowsIllegalStateException() {
-                // locationIndex non initialisé devrait lancer exception
-                assertThrows(IllegalStateException.class, () -> new GraphHopper().getLocationIndex(),
-                                "LocationIndex not initialized");
-        }
-
-        @Test
-        void getBaseGraphThrowsIllegalStateException() {
-                // baseGraph non initialisé devrait lancer exception
-                IllegalStateException exception = assertThrows(IllegalStateException.class,
-                                () -> new GraphHopper().getBaseGraph());
-                assertEquals("GraphHopper storage not initialized", exception.getMessage());
-        }
-
-        @Test
-        void setOSMFileThrowsIllegalArgumentException() {
-                // paramètre string vide devrait lancer exception
-                IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                                () -> new GraphHopper().setOSMFile(""));
-                assertEquals("OSM file cannot be empty.", exception.getMessage());
-        }
-
-        @Test
-        void setGraphHopperLocationThrowsIllegalArgumentException() {
-                // paramètre ghLocation null devrait lancer exception
-                IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                                () -> new GraphHopper().setGraphHopperLocation(null));
-                assertEquals("graphhopper location cannot be null", exception.getMessage());
-        }
-
 }
